render table with translation
