+++ frontend/src/entities/task/index.ts
@@ -7,3 +7,4 @@ export * from "./ui/task-type-badge.tsx";
 export * from "./ui/select-assignee";
 export * from "./ui/task-description.tsx";
 export * from "./ui/select-task-status.tsx";
+export * from "./ui/task-changes.tsx";
